+++ bundles/org.openhab.ui/web/src/components/widgets/system/index.js
@@ -14,3 +14,4 @@ export { default as OhPlayerControls } from './oh-player-controls.vue'
 export { default as OhList } from './oh-list.vue'
 export { default as OhSwiper } from './oh-swiper.vue'
 export { default as OhSwiperSlide } from './oh-swiper-slide.vue'
+export { default as OhTrend } from './oh-trend.vue'
